feat: update SubjectStat fields to use float32 for improved precision in statistics

+++ server/modules/stats/DTO/subject_stat.dto.go
@@ -4,8 +4,8 @@ import surrealModels "github.com/surrealdb/surrealdb.go/pkg/models"
 
 type SubjectStat struct {
 	Count      uint                   `json:"count"`
-	Difficulty uint                   `json:"difficulty"`
-	Grade      uint                   `json:"grade"`
-	Workload   uint                   `json:"workload"`
+	Difficulty float32                `json:"difficulty"`
+	Grade      float32                `json:"grade"`
+	Workload   float32                `json:"workload"`
 	Trimester  surrealModels.RecordID `json:"trimester" swaggertype:"object"`
 }
